requirements.txt
argon2-cffi>=21.3.0
cryptography>=38.0.0
requests>=2.28.0
pyperclip>=1.8.0